# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modern RSS reader API.",
        "title": "Gist API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/settings/ai": {
            "get": {
                "description": "Get the AI provider configuration with masked API keys",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get AI settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.aiSettingsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the AI provider configuration. Empty apiKey keeps existing key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update AI settings",
                "parameters": [
                    {
                        "description": "AI settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.aiSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.aiSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/settings/ai/test": {
            "post": {
                "description": "Test the AI provider connection with a \"Hello world\" message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Test AI connection",
                "parameters": [
                    {
                        "description": "AI test configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.aiTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.aiTestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.aiSettingsRequest": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "autoTranslate": {
                    "type": "boolean"
                },
                "baseUrl": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "reasoningEffort": {
                    "type": "string"
                },
                "summaryLanguage": {
                    "type": "string"
                },
                "thinking": {
                    "type": "boolean"
                },
                "thinkingBudget": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.aiSettingsResponse": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "autoTranslate": {
                    "type": "boolean"
                },
                "baseUrl": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "reasoningEffort": {
                    "type": "string"
                },
                "summaryLanguage": {
                    "type": "string"
                },
                "thinking": {
                    "type": "boolean"
                },
                "thinkingBudget": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.aiTestRequest": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "baseUrl": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "reasoningEffort": {
                    "type": "string"
                },
                "thinking": {
                    "type": "boolean"
                },
                "thinkingBudget": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.aiTestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}